Move viewPagerAdapter to adapter package
